Add handling multiple pressed buttons on Deco 03
The current handling of pad buttons is taking into
account only the last pressed button and considering
everything is unpressed when any button is.
This change treats each button press/unpress separatly.

I do not know if the current behaviour is intended or not,
but this one seems to make more sense (and it is the behaviour
on the vendor driver for the Deco 03).
Anyway, this commit is just a quick implementation of the behaviour,
and only implemented on the deco_03 level, TODO: make an issue and
ask to make the change globally and make a push request