Have Pen<-->Brush switch made with BTN_STYLUS2
Current code expects fourth bit in data[1] to indicate
whether the pen is in brush mode or not.
For the Deco 03, this bit is always set to 0, and with
the vendor driver the switch is made by pressing the second
stylus button.
This is a quick and dirty implementation to have this behaviour.

I do not know the behaviour in other tablets, so this can
only be done on the deco_03 level. This implementation is done
on the xp_pen_unified_device level and so should be changed if
to be integrated.(TODO)